refactor: change to new CustomJwtException
